+++ tests/TestCase/Processor/Image/ImageProcessorTest.php
@@ -15,6 +15,7 @@
 namespace PhpCollective\Test\TestCase\Processor\Image;
 
 use Intervention\Image\Drivers\Gd\Driver;
+use Intervention\Image\Drivers\Imagick\Driver as ImagickDriver;
 use Intervention\Image\ImageManager;
 use PhpCollective\Infrastructure\Storage\File;
 use PhpCollective\Infrastructure\Storage\FileFactory;
@@ -34,8 +35,7 @@ class ImageProcessorTest extends TestCase
      */
     public function testProcessor(): void
     {
-        $fileStorage = $this->getMockBuilder(FileStorageInterface::class)
-            ->getMock();
+        $fileStorage = $this->createStub(FileStorageInterface::class);
 
         $pathBuilder = new PathBuilder();
 
@@ -64,6 +64,78 @@ class ImageProcessorTest extends TestCase
 
         $file = $file->withVariants($collection->toArray());
 
+        $file = $processor->process($file);
+
+        $this->assertInstanceOf(File::class, $file);
+    }
+
+    /**
+     * @return void
+     */
+    public function testProcessorAppliesRepeatedOperationsInOrder(): void
+    {
+        $applied = [];
+        $fileStorage = $this->createStub(FileStorageInterface::class);
+
+        $processor = new ImageProcessor(
+            $fileStorage,
+            new PathBuilder(),
+            new ImageManager(new Driver()),
+        );
+
+        $file = FileFactory::fromDisk($this->getFixtureFile('titus.jpg'), 'local')
+            ->withUuid('914e1512-9153-4253-a81e-7ee2edc1d973')
+            ->withFilename('foobar.jpg')
+            ->addToCollection('avatar')
+            ->belongsToModel('User', '1');
+
+        $collection = ImageVariantCollection::create();
+        $collection
+            ->addNew('effects')
+            ->callback(function ($image, $args) use (&$applied): void {
+                $applied[] = 'first';
+            })
+            ->callback(function ($image, $args) use (&$applied): void {
+                $applied[] = 'second';
+            });
+
+        $file = $file->withVariants($collection->toArray());
+        $processor->process($file);
+
+        $this->assertSame(['first', 'second'], $applied);
+    }
+
+    /**
+     * @return void
+     */
+    public function testProcessorWithImagickDriver(): void
+    {
+        if (!extension_loaded('imagick')) {
+            $this->markTestSkipped('Imagick driver requires the Imagick PHP extension');
+        }
+
+        $fileStorage = $this->createStub(FileStorageInterface::class);
+
+        $processor = new ImageProcessor(
+            $fileStorage,
+            new PathBuilder(),
+            new ImageManager(new ImagickDriver()),
+        );
+
+        $file = FileFactory::fromDisk($this->getFixtureFile('titus.jpg'), 'local')
+            ->withUuid('914e1512-9153-4253-a81e-7ee2edc1d973')
+            ->withFilename('foobar.jpg')
+            ->addToCollection('avatar')
+            ->belongsToModel('User', '1');
+
+        $collection = ImageVariantCollection::create();
+        $collection
+            ->addNew('resizeAndFlip')
+            ->flipHorizontal()
+            ->resize(300, 300)
+            ->optimize();
+
+        $file = $file->withVariants($collection->toArray());
         $file = $processor->process($file);
 
         $this->assertInstanceOf(File::class, $file);